kernel: prevent the motor to move during the first MOTOR_USE_WAIT_MS ms after device is booted
Somehow when the device is booting selinux enforced, the xiaomi motor
service is sending the kernel driver the command to move up several
times. Logging does not show any information why this is happening.
When Booting selinux permissive, this is not happening.

Until the real reason is found, this dirty hack is used to be able to
boot selinux enforced without the camera popping out during boot.

Signed-off-by: Millan Mishra <3ztremors@gmail.com>